Overhaul asset pipeline: scene assets, registry, instantiation, nlohmann JSON
- Rename XJGltfImporter to XJModelImporter with static ExtractMesh
- Add XJAssetRef: lightweight Handle+Type pair with URI support
- Add XJAssetRegistry: in-memory asset database with JSON save/load persistence
- Add XJSceneAsset: disk-side scene data with entities, transforms, meshes, cameras, lights
- Add XJSceneInstantiator: converts XJSceneAsset into live ECS entities with hierarchy
- Add XJSceneAssetSerializer: reads/writes .xjscene files as JSON
- Add ECS components: XJSceneAssetRefComponent, XJMeshAssetRefComponent, XJMaterialAssetRefComponent
- Add nlohmann/json v3.12.0 as external JSON library dependency
- ECS improvements: const-correctness for HasComponent/GetComponent, UUID widened to uint64_t, XJNode parent management fixes
- Refactor main.cpp to asset-driven pipeline: registry, scene asset, instantiator
- Add Resource/Config/AssetRegistry.json and Resource/Scenes/Default.xjscene
- CMake: add json library, GCC big-obj flag, post-build copy for Scenes

+++ Core/Private/ECS/XJUUID.cpp
@@ -6,14 +6,14 @@ namespace XJ
     //随机引擎
     static std::random_device s_RandomDevice;
     static std::mt19937_64 s_Engine(s_RandomDevice());
-    static std::uniform_int_distribution<uint32_t> s_UniformDistribution(1, UINT32_MAX);
+    static std::uniform_int_distribution<uint64_t> s_UniformDistribution(1, UINT64_MAX);
 
 
     XJUUID::XJUUID(/* args */) : mUUID(s_UniformDistribution(s_Engine))
     {
 
     }
-    XJUUID::XJUUID(uint32_t uuid) : mUUID(uuid)
+    XJUUID::XJUUID(uint64_t uuid) : mUUID(uuid)
     {
 
     }
